Serial number and part number support
Add the feature to set a serial number and a part number for the device. Sai numbers are then sent to Astarte and stored in the NVS. To ease the nvs use, now `edgehog_new`takes an `edgehog_device_config_t`designed to centralize all the parameters needed to instantiate an `edgehog_device`. At the moment `edgehog_device_config_t` takes a required `astarte_device_handle_t` and an optional partition label where the numbers are stored.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>